Calculate a student's average
Read four numbers (N1, N2, N3, N4), each with one decimal place, corresponding to the four grades of a student. Calculate the average with weights of 2, 3, 4 and 1, respectively, for each of these grades and display this average accompanied by the message "Average: ". If this average is greater than or equal to 7.0, print the message "Student passed.". If the calculated average is less than 5.0, print the message "Student Failed.". If the calculated average is a value between 5.0 and 6.9, including these, the program should print the message "Student in exam.".

In case the student is in the exam, read a value corresponding to the exam grade obtained by the student. Then print the message "Exam grade: " accompanied by the grade typed. Recalculate the average (add the exam score to the previously calculated average and divide by 2). and print the message "Student passed." (if the final average is 5.0 or more) or "Failed student.", (if the average is 4.9 or less). For these two cases (passed or failed after taking the exam) present in the last line a message "Final average: " followed by the final average for that student.