Coordinates of a Point in the Quadrants
Read 2 values with one decimal place (x and y), which should represent the coordinates of a point on a plane. Next, determine which quadrant the point belongs to, or whether it is on one of the Cartesian axes or at the origin (x = y = 0).

If the point is at the origin, write the message “Origin”.

If the point is on one of the axes, write “X Axis” or “Y Axis”, depending on the situation.