Triangle classification
Read 3 floating point values A, B and C and sort them in descending order so that side A represents the longest of the 3 sides. Next, determine the type of triangle these three sides form, based on the following cases, always writing a suitable message:
if A ≥ B+C, display the message: DOES NOT FORM TRIANGLE
if A2 = B2 + C2, display the message: RECTANGLE TRIANGLE
if A2 > B2 + C2, display the message: TRIANGLE OBTUSANGLE
if A2 < B2 + C2, display the message: SHARP TRIANGLE
if all three sides are equal, display the message: EQUILATERAL TRIANGLE
if only two sides are the same, display the message: ISOSCELES TRIANGLE
Input
The input contains three double-precision floating point values A (0 < A), B (0 < B) and C (0 < C).

Exit
Print all ratings for the triangle specified in the input.