Time game
Read the start time and end time of a game. Then calculate the duration of the game, knowing that it can start on one day and end on another, with a minimum duration of 1 hour and a maximum of 24 hours.

Input
The input contains two integer values representing the start time and end time of the game.

Exit
Show the game duration as shown below.