Conditional Structure for selection
Read an integer that represents a long distance dialing code. Then, inform which city the area code belongs to, considering the table below:

If the input is any other DDD that is not present in the table above, the program should inform:
Area code not registered

Input
The input consists of a single integer value.

Exit
Print the city name corresponding to the existing area code in the entry. Print unregistered area code if there is no area code corresponding to the number entered.