Selection Structure
In this problem, you must read 3 words that define the possible type of animal according to the diagram below, from left to right. Then conclude which of the following animals was chosen, using the three words provided.

Input
The entry contains 3 words, one on each line, necessary to identify the animal according to the figure above, with all lowercase letters.

Exit
Print the name of the animal corresponding to the entry provided.