Automate calculation of the duration of an event
Pedrinho is organizing an event at his University. The event should be in the month of April, starting and ending within the month. The problem is that Pedrinho wants to calculate how long the event will last, since he knows when the event starts and when it ends.

Knowing that the event can last from a few seconds to several days, you should help Pedrinho calculate the duration of this event.

Input
As an input, the first line will contain the description “Day”, followed by a space and the day of the month on which the event will start. In the next line, the moment when the event will start will be informed, in the format hh : mm : ss. In the third and fourth input lines there will be other information in the same format as the first two lines, indicating the end of the event.

Exit
At the output, the duration of the event must be presented, in the following format:

W day(s)
X hour(s)
Y minute(s)
Z second(s)

Note: Consider that the test case event for the problem has a minimum duration of 1 minute.