Sum of Consecutive Odds I
Read 2 integer values X and Y. Then calculate and display the sum of the odd numbers between them.

Input
The input file contains two integer values.

Exit
The program must print an integer value. This value is the sum of the odd values that are among the values given in the input that must fit in an integer.