Filter if it's Odd or Even
Read an integer value N. This value will be the number of values that will be read next. For each read value, display a message in English saying whether this read value is even (EVEN), odd (ODD), positive (POSITIVE) or negative (NEGATIVE). In case the value is equal to zero (0), although the correct description is (EVEN NULL), because by definition zero is even, your program should print only NULL.

Input
The first line of the input contains an integer value N(N < 10000) that indicates the number of test cases. Each test case below is an integer value X (-107 < X < 107).

Exit
For each case, print a corresponding message, according to the example below. All letters must be capitalized and there must always be a space between two words printed on the same line.