Weighted average
Read 1 integer value N, which represents the number of test cases that follows. Each test case consists of 3 real values, each with one decimal place. Present the weighted average for each of these sets of 3 values, with the first value having a weight of 2, the second value having a weight of 3 and the third value having a weight of 5.

Input
The input file contains an integer value N on the first line. Each N line below contains a test case with three values with one decimal place each value.

Exit
For each test case, print the weighted average of the 3 values, as shown in the example below.