Countable experiences
Maria has just started her delivery course at the medical school and needs your help to organize the experiments of a laboratory she is responsible for. She wants to know at the end of the year, how many guinea pigs were used in the laboratory and the percentage of each type of guinea pig used.

This lab in particular uses three types of test subjects: frogs, mice and rabbits. To obtain this information, she knows exactly the number of experiments that were performed, the type of guinea pig used and the number of guinea pigs used in each experiment.

Input
The first line of input contains an integer value N that indicates the various test cases that follow. Each test case contains an integer Amount (1 ≤ Amount ≤ 15) that represents the amount of guinea pigs used and a Type character ('C', 'R' or 'S'), indicating the type of guinea pig (R:Rato S :Toad C:Rabbit).

Exit
Present the total number of guinea pigs used, the total of each type of guinea pig used and the percentage of each one in relation to the total number of guinea pigs used, and the percentage must be presented with two digits after the dot.