Numbers sequence and sum them
Read an undetermined set of M and N value pairs (stop when either value is less than or equal to zero). For each pair read, show the sequence from smallest to largest and the sum of consecutive integers between them (including N and M).

Input
The input file contains an unspecified number of M and N values. The last input line will contain a null or negative number.

Exit
For each pair of values, print the sequence from the smallest to the largest and their sum, as shown in the example below.