Sort by quadrant
Write a program to read the coordinates (X,Y) of an indeterminate number of points in the Cartesian system. For each point write the quadrant to which it belongs. The algorithm will be terminated when at least one of the two coordinates is NULL (in this situation without writing any message).

Input
The input contains several test cases. Each test case contains 2 integer values.

Exit
For each test case, show in which quadrant of the Cartesian system the read coordinate is found, as shown in the example.