Divide x by y
Write an algorithm that reads 2 numbers and prints the result of dividing the first by the second. If it is not possible, display the message “impossible division” for the values in question.

Input
The input contains an integer N. This N will be the number of pairs of integer values (X and Y) that will be read next.

Exit
For each case, show the result of the division with a digit after the decimal point, or “impossible division” if the calculation is not possible.