Grade evaluation
Write a program that reads the grades for a student's two assessments. Calculate and print the semester average. Make the algorithm only accept valid grades (a valid grade must belong to the range [0,10]). Each note must be validated separately.

Input
The input contains multiple real values, positive or negative. The program must be terminated when two valid notes are read.

Exit
If an invalid grade is read, the message "invalid grade" should be printed.
When two valid grades are read, the message "media = " must be printed followed by the calculation value. The value must be presented with two places after the decimal point.