Multiple Notes With Validation
Write a program to read a student's first and second assessment grades. Calculate and print the semester average. The program should only accept valid grades (a valid grade must belong to the range [0,10]). Each note must be validated separately.

At the end, the message “new calculation (1-yes 2-nao)” must be printed, asking the user to inform a code (1 or 2) indicating whether or not he wants to run the algorithm again, (accept only code 1 or two). If code 1 is informed, the execution of the entire program must be repeated to allow a new calculation, otherwise the program must be terminated.

Input
The input file contains multiple real values, positive or negative. When two valid notes are read, an integer value X must be read. The program must stop when the value read for this X is equal to 2.

Exit
If an invalid grade is read, the message “invalid grade” should be printed. When two valid grades are read, the message “average = ” must be printed followed by the calculation value.

Before reading X, the message "new calculation (1-yes 2-no)" must be printed and this message must be displayed again if the value of the standard input for X is less than 1 or greater than 2, according to the example below.

The mean must be printed with two digits after the decimal point.