Evaluate number of soccer games
The Gaucho Football Federation hired you to write a program to make a statistic of the results of several GRENAIS. Write a program to read the number of goals scored by Inter and Grêmio in a GRENAL. Right after writing the message "New grenal (1-yes 2-no)" and requesting a response. If the answer is 1, the algorithm must be run again asking for the number of goals scored by the teams in a new match, otherwise it must be terminated by printing:

- How many GRENAIS were part of the statistic.
- The number of Inter victories.
- The number of Grêmio victories.
- The number of Ties.
- A message indicating which team won the most GRENAIS (or "There was no winner", if it ends in a tie).

Input
The input file contains 2 integer values, corresponding to the goals scored by Inter and Grêmio respectively. Then there will be an integer (1 or 2), corresponding to the repetition of the program.

Exit
After each reading of the goals, the message "New grenal (1-yes 2-no)" must be printed. When the algorithm is terminated, the statistics must be shown as described above. Note: the word "Gremio" must be printed without an accent, as shown in the example below.