Type of fuel
A gas station wants to determine which of its products is preferred by its customers. Write an algorithm to read the type of fuel filled (coded as follows: 1.Alcohol 2.Gasoline 3.Diesel 4.End). If the user enters an invalid code (outside the range 1 to 4), a new code must be requested (until it is valid). The program will be terminated when the code entered is number 4.

Input
The input contains only integer and positive values.

Exit
The message must be written: "THANK YOU SO MUCH" and the number of customers who supplied each type of fuel, as shown in the example.