Adding consecutive integers
Make an algorithm to read an A value and an N value. Print the sum of A + i for each i with the values (0 <= i <= N-1). As long as N is negative or ZERO, a new N(only N) must be read.

Input
The input contains only integer values, which can be positive or negative. All values are on the same line.

Exit
The output contains only an integer value.