Conditional number sequence
Write a program that reads two integers: X and Z (as many values ​​for Z must be read as necessary, until a value greater than X is entered for it). Count how many integers must be added in sequence (considering the X in this sum) so that the sum exceeds Z as little as possible. Write the final count value.

The input can contain, for example, the values ​​21 21 15 30. In this case, the value 21 is assumed for X while the values ​​21 and 15 must be disregarded as they are less than or equal to X. As the value 30 is within the specification (greater than X) it will be valid and then the calculations must be processed to present the value 2 in the output, since it is the amount of values ​​added to produce a value greater than 30 (21 + 22).

Input
The input contains only integer values, one per line, which can be positive or negative. The first input value will be the value of X. The next input line will contain Z. If Z does not meet the problem specification, it must be read again as many times as necessary.

Exit
Print a line with an integer representing the number of integers that must be added, according to the specification above.