Fibonacci Sequence
The following sequence of numbers 0 1 1 2 3 5 8 13 21... is known as the Fibonacci series. In this sequence, each number after the first 2 is equal to the sum of the previous 2. Write an algorithm that reads an integer N (N < 46) and displays the first N numbers of this series.

Input
The input file contains an integer value N (0 < N < 46).

Exit
The values must be shown on the same line, separated by a blank space. There must be no space after the last value.

Input example
5

Output Example
0 1 1 2 3