Circle area
The formula for calculating the area of a circle is: area = π . radius2. Considering for this problem that π = 3.14159:

- Calculate the area by squaring the radius value and multiplying by π.

Input
The input contains a floating point value (double precision), in this case, the radius variable.

Exit
Display the message "A=" followed by the value of the variable area, as shown below, with 4 places after the decimal point. Use double precision variables. Like all problems, don't forget to print the end of line after the result, otherwise you'll get "Presentation Error".

Input Examples
2.00

Output Examples
A=12.5664