Average calculation
Read 2 double-precision floating point values A and B, which correspond to 2 grades of a student. Next, calculate the student's average, knowing that grade A has a weight of 3.5 and grade B has a weight of 7.5 (the sum of the weights is therefore 11). Assume that each grade can go from 0 to 10.0, always with one decimal place.

Input
The input file contains 2 values with one decimal place each.

Exit
Print the message "AVERAGE" and the student's average as shown below, with 5 digits after the decimal point and with a blank space before and after the equality. Use double precision variables and like all problems, don't forget to print the end of line after the result, otherwise you will get "Presentation Error".

Input Examples
5.0
7.1

Output Examples
AVERAGE = 6.43182