Calculation of the average 2
Read 3 values, in this case, variables A, B and C, which are the three grades of a student. Next, calculate the student's average, knowing that grade A has a weight of 2, grade B has a weight of 3 and grade C has a weight of 5. Consider that each grade can go from 0 to 10.0, always with one decimal place.

Input
The input file contains 3 values with one decimal place, double precision.

Exit
Print the message "AVERAGE" and the student's average as shown below, with 1 digit after the decimal point and with a blank space before and after the equality. As with all problems, don't forget to print the end of line after the result, otherwise you'll get "Presentation Error".

Input Examples
5.0
6.0
7.0

Output Examples
AVERAGE = 6.3