Salary calculation
Write a program that reads an employee's number, their number of hours worked, the hourly rate they earn, and calculates that employee's salary. Next, show the employee's number and salary, to two decimal places.

Input
The input file contains 2 whole numbers and 1 number with two decimal places, representing the number, amount of hours worked and the amount the employee receives per hour worked, respectively.

Exit
Print the employee's number and salary, as the example provided, with a blank space before and after the equality. In the case of salary, there must also be a blank space after the $.

Input Examples
25
100
5.50

Output Examples
NUMBER = 25
SALARY = U$ 550.00