Simple calculation
In this problem, you must read the code of a part 1, the number of parts 1, the unit value of each part 1, the code of a part 2, the number of parts 2 and the unit value of each part 2. After , calculate and display the amount to be paid.

Input
The input file contains two lines of data. In each line there will be 3 values, respectively two integers and a value with 2 decimal places.

Exit
The output should be a message as shown below, remembering to leave a space after the colon and a space after the "R$". The value must be presented with 2 places after the dot.

Input Examples
12 1 5.30
16 2 5.10

Output Examples
AMOUNT PAYABLE: BRL 15.50