Calculate the area of the sphere
Make a calculation program and most of the volume of a sphere being represented by the value of its radius. The formula for calculating volume is: (4/3) * pi * R3. Consider (assign) to pi the value 3.14159.

Tip: When using a formula, try to use (4/3.0) or (4.0/3), as some languages (including C++) assume that the result of dividing two integers is another integer.

Input
The input file contains a floating point (double precision) value, corresponding to the radius of the sphere.

Exit
The output should be a message as in the example below and below, with a space after the equality. The value must be presented with 3 places after the dot.

Input Examples
3

Output Examples
VOLUME = 113.097