Calculation of areas of △, ○, trapezoid, ▀ and ▂
Write a program that reads three double-precision floating-point values: A, B, and C. Then calculate and display:
a) The area of a right triangle that has A for its base and C for its height.
b) the area of the circle of radius C. (pi = 3.14159)
c) the area of the trapezoid that has A and B for bases and C for height.
d) the area of the square that has side B.
e) the area of the rectangle that has sides A and B.
Input
The input file contains three values with one digit after the decimal point.

Exit
The output file should contain 5 lines of data. Each line corresponds to one of the areas described above, always with a corresponding message and a space between the colon and the value. The calculated value must be presented with 3 digits after the decimal point.

Input Examples
3.0 4.0 5.2

Output Examples
TRIANGLE: 7,800
CIRCLE: 84,949
TRAPEZIUM: 18,200
SQUARE: 16,000
RECTANGLE: 12,000