Fuel expense
Joaozinho wants to calculate and show the amount of liters of fuel spent in a trip, when using a car that makes 12 KM/L. For this, he would like you to help him through a simple program. To make the calculation, you must provide the time spent on the trip (in hours) and the average speed during the trip (in km/h). Thus, one can obtain distance traveled and then calculate how many liters would be needed. Show the value to 3 decimal places after the dot.

Input
The input file contains two integers. The first is the time spent on the trip (in hours) and the second is the average speed during the trip (in km/h).

Exit
Print the number of liters needed to complete the trip, with three digits after the decimal point

Input example
10
85

Output Example
70,833