Number of banknotes
Read an integer value. Next, calculate the smallest possible number of bills (ballots) into which the value can be decomposed. The grades considered are 100, 50, 20, 10, 5, 2 and 1. Then show the value read and the list of grades required.

Input
The input file contains an integer value N (0 < N < 1000000).

Exit
Print the value read and then the minimum amount of notes of each type required, as shown in the example provided. Do not forget to print the end of line after each line, otherwise your program will display the message: “Presentation Error”.

Input example
576

Output Example
576
5 bill(s) of BRL 100.00
1 bill(s) of BRL 50.00
1 bill(s) of BRL 20.00
0 bill(s) of BRL 10.00
1 bill(s) of BRL 5.00
0 bill(s) of BRL 2.00
1 bill(s) of BRL 1.00