Calculate the average age
Write an algorithm to read an indeterminate number of data, each containing the age of an individual. The last data, which will not be included in the calculations, contains the negative age value. Calculate and print the average age of this group of individuals.

Input
The input contains an unspecified number of integers. The input will terminate when a negative value is read.

Exit
The output contains a value corresponding to the average age of the individuals.

The mean must be printed with two digits after the decimal point.

Input example
34
56
44
23
-two

Output Example
39.25