Sum of consecutive odd numbers III
Read an integer value N which is the number of test cases that follow. Each test case consists of two integers X and Y. You must return the sum of consecutive odd Ys starting from X including X itself if it is odd. For example:
for input 4 5, the output should be 45, which is equivalent to: 5 + 7 + 9 + 11 + 13
for input 7 4, the output should be 40, which is equivalent to: 7 + 9 + 11 + 13

Input
The first line of input is an integer N which is the number of test cases that follow. Each test case consists of a line containing two integers X and Y.

Exit
Print the sum of consecutive odd numbers starting from the X value.

Input example
two
4 3
11 2

Output example
21
24