Sum of consecutive pairs
The program must read an integer value X indefinitely. (The program will stop when the value of X equals 0). For each X read, print the sum of the 5 consecutive pairs starting from X, including the X , if even. If the input value is 4, for example, the output must be 40, which is the result of the operation: 4+6+8+10+12, while if the input value is 11, for example, the output must be 80, which is the sum of 12+14+16+18+20.

Input
The input file contains many integer values. The last value in the file is zero.

Exit
Print the output as explained above and the example below.

Input example
4
11
0

Output example
40
80