Calculation of population growth rate
Mariazinha wants to solve an interesting problem. Given the population information and the growth rate of any two cities (A and B), she would like to know how many years it will take for the smaller city (always city A) to surpass city B in population. Of course it only wants to know for cities whose growth rate of city A is higher than the growth rate of city B, so it has previously sorted out for you only the test cases that have the highest growth rate for city A. Your task is to build a program that displays the time in years for each test case.

However, in some cases the time can be very long, and Mariazinha is not interested in knowing the exact time for these cases. You only need to inform, in this situation, the message "More than 1 century.".

Input
The first line of the input contains a single integer T, indicating the number of test cases (1 ≤ T ≤ 3000). Each test case contains 4 numbers: two integers PA and PB (100 ≤ PA < 1000000, PA < PB ≤ 1000000) respectively indicating the population of A and B, and two values ​​G1 and G2 (0.1 ≤ G1 ≤ 10.0, 0.0 ≤ G2 ≤ 10.0, G2 < G1) with one digit after the decimal point each, respectively indicating the population growth of A and B (in percentage).

Attention: Population is always an integer value, so a 2.5% growth over a population of 100 people will result in 102 people, not 102.5 people, while a 2.5% growth over a population of 1000 people will result in 1025 people. Also, do not use single-precision variables for growth rates.

Exit
Print, for each test case, how many years it will take for city A to surpass city B in number of inhabitants. Note: if the time is more than 100 years, the program must display the message: More than 1 century. In this case, I believe that it is better to stop the program immediately after passing 100 years, otherwise you may receive "Time Limit Exceeded" as a response from the submission of this problem.

Input example
6
100 150 1.0 0
90000 120000 5.5 3.5
56700 72000 5.2 3.0
123 2000 3.0 2.0
100000 110000 1.5 0.5
62422 484317 3.1 1.0

Output Example
51 years.
16 years.
12 years.
More than 1 century.
10 years.
100 years.