Evaluate whether a number is perfect.
In mathematics, a perfect number is an integer for which the sum of all its positive proper divisors (excluding itself) is equal to the number itself. For example the number 6 is perfect, because 1+2+3 is equal to 6. Your task is to write a program that prints whether a certain number is perfect or not.

Input
The input contains several test cases. The first line of the input contains an integer N (1 ≤ N ≤ 20), indicating the number of test cases in the input. Each of the next N lines contains an integer value X (1 ≤ X ≤ 108), which may or may not be a perfect number.

Exit
For each input test case, print the message “X is perfect” or “X is not perfect”, according to the given specification.

Input example
3
6
5
28

Output example
6 is perfect
5 is not perfect
28 is perfect