Prime number
In mathematics, a Prime Number is one that can be divided only by 1 (one) and itself. For example, the number 7 is prime as it can only be divided by the number 1 and the number 7.

Input
The input contains several test cases. The first line of the input contains an integer N (1 ≤ N ≤ 100), indicating the number of test cases in the input. Each of the next N lines contains an integer value X (1 < X ≤ 107), which may or may not be a prime number.

Exit
For each input test case, print the message “X is prime” or “X is not prime”, according to the given specification.

Input example
3
8
51
7

Output example
8 is not cousin
51 is not cousin
7 is cousin