Vector Filling I
Read a value and write a program that places the read value in the first position of an N[10] vector. In each subsequent position, put twice the value of the previous position. For example, if the value read is 1, the values of the vector must be 1,2,4,8 and so on. Show the vector next.

Input
The input contains an integer value (V<=50).

Exit
For each position of the vector, write "N[i] = X", where i is the position of the vector and X is the value stored in position i. The first number of the vector N (N[0]) will receive the value of V.

Input example
1

Output example
N[0] = 1
N[1] = 2
N[2] = 4
...