Exchange in Vector I
Write a program that reads an N[20] vector. Then replace the first element with the last, the second element with the second to last, etc., until the 10th is replaced with the 11th. Show the modified vector.

Input
The input contains 20 integer values, positive or negative.

Exit
For each position of the vector N, write "N[i] = Y", where i is the position of the vector and Y is the value stored at that position.

Input example
0
-5
...
63
230


Output example
N[0] = 230
N[1] = 63
...
N[18] = -5
N[19] = 0