Fibonacci in Vector
Write a program that reads a value and displays the Fibonacci number corresponding to this read value. Remember that the first 2 elements of the Fibonacci series are 0 and 1 and each next term is the sum of the 2 before it. All Fibonacci values calculated in this problem must fit in an unsigned 64-bit integer.

Input
The first line of input contains an integer T, indicating the number of test cases. Each test case contains a single integer N (0 ≤ N ≤ 60), corresponding to the N-th term of the Fibonacci series.

Exit
For each test case in the input, print the message "Fib(N) = X", where X is the Nth term of the Fibonacci series.

Input example
3
0
4
two

Output example
Fib(0) = 0
Fib(4) = 3
Fib(2) = 1