Vector Filling III
Read an X value. Place this value in the first position of an N[100] vector. In each subsequent position from N (1 to 99), place half of the value stored in the previous position, as shown in the example below. Print the N vector.

Input
The input contains a double precision value to 4 decimal places.

Exit
For each position of the vector N, write "N[i] = Y", where i is the position of the vector and Y is the value stored at that position. Each vector value must be presented to 4 decimal places.

Input example
200,0000

Output example
N[0] = 200,0000
N[1] = 100,0000
N[2] = 50,0000
N[3] = 25,0000
N[4] = 12,5000
...