Vector Fill IV
In this problem you must read 15 values and place them in 2 vectors depending on whether these values are even or odd. But the size of each of the two vectors is 5 positions. Then, each time one of the two vectors fills up, you must print the entire vector and use it again for the next numbers that are read. After reading, print the remaining content in each of the two vectors, first printing the values of the odd vector. Each vector can be filled in as many times as needed.

Input
The input contains 15 integers.

Exit
Print the output as shown in the example below.

Input example
1
3
4
-4
two
3
8
two
5
-7
54
76
789
23
98

Output example
par[0] = 4
par[1] = -4
par[2] = 2
par[3] = 8
par[4] = 2
odd[0] = 1
odd[1] = 3
odd[2] = 3
odd[3] = 5
odd[4] = -7
odd[0] = 789
odd[1] = 23
par[0] = 54
par[1] = 76
par[2] = 98