Minor and Position
Input
The first input line contains a single integer N (1 < N < 1000), indicating the number of elements that should be read next into the X[N] array of integers. The second line contains each of the N values, separated by a space. It is worth remembering that no entries will have repeated numbers.

Exit
The first line displays the message “Lowest value:” followed by a space and the smallest value read in the input. The second line displays the message “Posicao:” followed by a space and the position of the vector in which the smallest value read is found, remembering that the vector starts at position zero.

Input example
10
1 2 3 4 -5 6 7 8 9 10

Output example
Lowest value: -5
position: 4