Line in Matrix
In this problem you must read a number, indicating a row of the matrix on which an operation must be performed, an uppercase character, indicating the operation that will be performed, and all elements of a matrix M[12][12]. Then calculate and display the sum or average of the elements that are in the green area of ​​the matrix, as appropriate. The image below illustrates the case of entering the value 2 for the matrix row, demonstrating the elements that must be considered in the operation.




Input
The first input line contains a number L (0 ≤ L ≤ 11) indicating the line that will be considered for operation. The second input line contains a single Uppercase T character ('S' or 'M'), indicating the operation (Sum or Average) to be performed on the matrix elements. The 144 floating point values ​​that make up the matrix follow, and it is filled line by line, from line 0 to line 11, always from left to right.

Exit
Print the requested result (the sum or average), with 1 place after the decimal point.

Input example
two
s
0.0
-3.5
2.5
4.1
...

Output example
12.6