Column in Matrix
In this problem you must read a number that indicates a column of a matrix on which an operation must be performed, an uppercase character, indicating the operation that will be performed, and all elements of a matrix M[12][12]. Then calculate and display the sum or average of the elements that are in the green area of ​​the matrix, as appropriate. The image below illustrates the case of entering the value 5 for the matrix column, demonstrating the elements that must be considered in the operation.




Input
The first input line contains a number C (0 ≤ C ≤ 11) indicating the column that will be considered for operation. The second input line contains a single Uppercase T character ('S' or 'M'), indicating the operation (Sum or Average) to be performed on the matrix elements. Following are the 144 floating point values ​​that make up the array.

Exit
Print the requested result (the sum or average), with 1 place after the decimal point.

Input example
5
s
0.0
-3.5
2.5
4.1
...

Output example
12.6