Building houses
Mr PI is a very famous builder in the city of Programolândia. He needs your help to find the best land in the city, in order to carry out the construction of the various house projects he owns.

Consider, for example, that he has a project to build a house measuring 8 meters by 10 meters, and the municipal legislation allows the construction of a maximum of 100% of the land. As all land in this city is perfectly square and the value of the sides of the house is just a reference for the total area to be built (80 square meters), Mr PI would need a land of 8,994 meters, which simplified would give as a result 8 meters and the actual size of the house would be 64 square meters. If the legislation allowed to use 50% of the land, it would have to have 160 meters so that 50% of it would be 80 square meters, enough for a house of 8 x 8 meters (64 square meters). In the first test case, as the percentage to build is only 20%, the land would have to be 20 meters on a side for 1/5 of this land to have the size of 80 square meters. Help Mr PI determine the minimum plot size.

Input
The input is composed of several test cases. Each test case is composed of three integers A, B and C ( > 0 and ≤ 1000) separated by a space. These numbers represent the measurements of the house (A and B) and the maximum percentage allowed to build in that neighborhood (C). A single value of 0 indicates the end of entries.

Exit
You must enter an integer number, which represents the measurement of the side of the land. This value should be truncated if necessary.

Input example
8 10 20
1 10 100
10 3 100
0

Output Example
20
3
5