The Slug Race
Slug racing is a sport that has grown a lot in recent years, causing many people to dedicate their lives trying to catch fast slugs, and train them to make millions in races around the world. However, the task of catching fast slugs is not a very easy task, as practically all slugs are very slow. Each slug is classified into a level depending on its speed:


Level 1: If the speed is less than 10 cm/h.
Level 2: If the speed is greater than or equal to 10 cm/h and less than 20 cm/h.
Level 3: If the speed is greater than or equal to 20 cm/h.


Your task is to identify which speed level of the fastest slug in a group of slugs.

Input
The input consists of multiple test cases, each consisting of two lines: The first line contains an integer L (1 ≤ L ≤ 500) representing the number of slugs in the group, and the second line contains L integers Vi (1 ≤ Vi ≤ 50) representing the velocities of each slug in the group.

The entry ends with the end of file (EOF).

Exit
For each test case, print a single line indicating the speed level of the fastest slug in the group.

Input example
10
10 10 10 10 15 18 20 15 11 10
10
1 5 2 9 5 5 8 4 4 3
10
19 9 1 4 5 8 6 11 9 7

Output example
3
1
two