Bazinga! The big bang theory.
In the eighth episode of the second season of The Big Bang Theory, The Lizard-Spock Expansion, Sheldon and Raj discuss which of the two is better: the Saturn 3 movie or the Deep Space 9 series. Raj's suggestion for resolving the impasse It's a Rock-Paper-Scissors contest. However, Sheldon argues that, if the parties involved know each other, between 75% and 80% of Rock-Paper-Scissors disputes end in a tie, and so suggests Rock-Paper-Scissors-Lizard-Spock.

The proposed game rules are:

the scissors cut the paper;
paper wraps stone;
the stone crushes the lizard;
the lizard poisons Spock;
Spock destroys the scissors;
the scissors decapitate the lizard;
the lizard eats the paper;
the paper contests Spock;
Spock vaporizes the stone;
the stone breaks the scissors.
Although the situation is not resolved in the episode (both choose Spock, resulting in a tie), it is not difficult to deduce what would happen if the dispute continued. If Sheldon won, he would revel in the victory, exclaiming "Bazinga!"; if Raj won, he would conclude that "Raj cheated!"; if the result was a tie, he would demand a new game: "Again!". Once the characters in the game chosen by both are known, make a program that prints Sheldon's probable reaction.

Input
The input consists of a series of test cases. The first line contains a positive integer T (T ≤ 100), which represents the number of test cases. Each test case is represented by a line of input, containing the choices of Sheldon and Raj, respectively, separated by a blank space. The possible choices are the characters in the game: rock, paper, scissors, lizard and Spock.

Output
For each test case, the message "Case #t: R" should be printed, where t is the number of the test case (whose count starts at number one) and R is one of the three possible Sheldon reactions: "Bazinga! ", "Raj cheated!", or "Again!".

Input Examples
3

stone paper

scissors lizard

Spock Spock

Output Examples

Case #1: Bazinga!

Case #2: Raj cheated!

Case #3: Again!